Added exception handling
Added more options for Z order positioning of windows.
   It is now possible to set XY and Z position in Windows message to gain speed.

Added support for mouse wheels.

Added support for HID (human Interface Devices, like joysticks).

When switching Visual ITEC forms, sometimes the background flashes in front. 
This was due the fact that the modal message loop of GWindows disables the 
parent window (in out case Visual ITEC). When the focus was lost, 
the next main window was selected for the focus. The solution was to 
enable the parent again before it looses the focus in the WM_QUIT/WM_CLOSE message.

When a window (or control) is created outside the GWindows application
(like extending a standard Windows dialog) 
and you try to attach it the regular way,
strange things happen.
For these cases it is made possible to link a window (or control) to GWindows.